// apps/backend/src/modules/email/enums/email.enums.ts
export enum EmailTemplate {
  EMAIL_VERIFICATION = 'EMAIL_VERIFICATION',
  PASSWORD_RESET = 'PASSWORD_RESET',
  PASSWORD_CHANGED = 'PASSWORD_CHANGED',
  ORDER_CONFIRMATION = 'ORDER_CONFIRMATION',
  ORDER_STATUS_UPDATE = 'ORDER_STATUS_UPDATE',
  PAYMENT_RECEIPT = 'PAYMENT_RECEIPT',
  REFUND_PROCESSED = 'REFUND_PROCESSED',
  RETURN_UPDATE = 'RETURN_UPDATE',
}

export enum EmailStatus {
  QUEUED = 'QUEUED',
  SENDING = 'SENDING',
  SENT = 'SENT',
  FAILED = 'FAILED',
  DEAD = 'DEAD',
}